Make Server directory. Stop all gameIDs from being valid.
